bots/telegram/requirements.txt
python-telegram-bot==21.5
aiohttp==3.10.5
python-dotenv==1.0.1
coinbase-advanced-py>=1.2.0
websockets>=12.0
python-socketio[client,asyncio_server]==5.11.0